# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FormaNT - Car Rental",
    "description": "Projeto para aluguel de carros desenvolvido pela turma de Java do FormaNT",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "Client",
      "description": "Operations about client"
    },
    {
      "name": "Employee",
      "description": "Operations about employee"
    }
  ],
  "paths": {
    "/v1/client": {
      "post": {
        "tags": [
          "Client"
        ],
        "summary": "Create client",
        "description": "Add a client to database",
        "operationId": "createClient",
        "requestBody": {
          "description": "Created Client Object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiClientResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ApiClientResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Get user by user name",
        "description": "",
        "operationId": "getUserByName",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Client"
        ],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Carlitos Tevez"
          },
          "cpfCnpj": {
            "type": "string",
            "example": "435.763.342-22"
          },
          "cnh": {
            "type": "string",
            "example": "874658935467"
          },
          "age": {
            "type": "integer",
            "example": 23
          },
          "telephone": {
            "type": "string",
            "example": "26995648354"
          },
          "emergencyContact": {
            "type": "string",
            "example": "26975648354"
          },
          "email": {
            "type": "string",
            "example": "carlitosboca@gmail.com"
          }
        },
        "xml": {
          "name": "client"
        }
      },
      "ApiClientResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "format": "uuid"
          }
        },
        "xml": {
          "name": "##default"
        }
      }
    },
    "requestBodies": {
      "Pet": {
        "description": "Pet object that needs to be added to the store",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Client"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Client"
            }
          }
        }
      },
      "UserArray": {
        "description": "List of user object",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Client"
              }
              
            }
          }
        }
      }
    }
  }
  
}